requirements.txt
requests==2.32.3
jinja2==3.1.4
pyusb==1.2.1